api/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
redis
prometheus-fastapi-instrumentator
httpx
pytest
pytest-asyncio
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-logging
